+++ packages/sdk/src/index.ts
@@ -2,10 +2,9 @@ export * from "./utils";
 export * from "./primitives";
 export * from "./proof";
 export * from "./crypto";
-export * from "./db";
+export * from "./store";
 export * from "./sync";
 export * from "./merkleProver";
-export * from "./notesManager";
 
 export {
   OperationRequest,
@@ -16,3 +15,5 @@ export { NocturneContext } from "./NocturneContext";
 export { OpPreparer } from "./opPreparer";
 export { OpSigner } from "./opSigner";
 export { OpProver } from "./opProver";
+export { NocturneDB } from "./NocturneDB";
+export { NocturneSyncer } from "./NocturneSyncer";