+++ packages/sdk/src/merkleProver/inMemory.ts
@@ -1,63 +1,22 @@
-import { Wallet, Wallet__factory } from "@nocturne-xyz/contracts";
 import { MerkleProof } from "@zk-kit/incremental-merkle-tree";
-import { ethers } from "ethers";
-import { Address, NoteTrait, BinaryPoseidonTree } from "../primitives";
-import { MerkleDB } from "../db";
+import { BinaryPoseidonTree } from "../primitives";
 import { MerkleProver } from "./abstract";
-import { fetchInsertions } from "../sync";
-
-const DEFAULT_START_BLOCK = 0;
-const MERKLE_NEXT_BLOCK_TO_INDEX = "MERKLE_NEXT_BLOCK_TO_INDEX";
-
-export interface InMemoryMerkleProverOpts {
-  startBlock?: number;
-}
+import { assertOrErr } from "../utils";
 
 export class InMemoryMerkleProver extends MerkleProver {
   readonly localTree: BinaryPoseidonTree;
-  protected contract: Wallet;
-  protected provider: ethers.providers.Provider;
-  protected db: MerkleDB;
-  protected startBlock: number;
 
-  constructor(
-    walletContractAddress: Address,
-    provider: ethers.providers.Provider,
-    db: MerkleDB,
-    opts?: InMemoryMerkleProverOpts
-  ) {
+  constructor() {
     super();
 
     this.localTree = new BinaryPoseidonTree();
-    this.provider = provider;
-    this.contract = Wallet__factory.connect(
-      walletContractAddress,
-      this.provider
-    );
-    this.db = db;
-    this.startBlock = opts?.startBlock ?? DEFAULT_START_BLOCK;
-  }
-
-  static async fromDb(
-    merkleAddress: Address,
-    provider: ethers.providers.Provider,
-    db: MerkleDB,
-    opts?: InMemoryMerkleProverOpts
-  ): Promise<InMemoryMerkleProver> {
-    const self = new InMemoryMerkleProver(merkleAddress, provider, db, opts);
-
-    for await (const leaf of db.iterLeaves()) {
-      self.localTree.insert(leaf);
-    }
-
-    return self;
   }
 
   root(): bigint {
     return this.localTree.root();
   }
 
-  count(): number {
+  async count(): Promise<number> {
     return this.localTree.count;
   }
 
@@ -65,33 +24,12 @@ export class InMemoryMerkleProver extends MerkleProver {
     return this.localTree.getProof(index);
   }
 
-  async fetchLeavesAndUpdate(): Promise<void> {
-    // TODO: load default from network-specific config
-    const nextBlockToIndex =
-      (await this.db.kv.getNumber(MERKLE_NEXT_BLOCK_TO_INDEX)) ??
-      this.startBlock;
-    const latestBlock = await this.provider.getBlockNumber();
-
-    const newLeaves = await this.fetchNewLeaves(nextBlockToIndex, latestBlock);
-
-    for (const leaf of newLeaves) {
-      await this.db.storeLeaf(this.localTree.count, leaf);
-      this.localTree.insert(leaf);
+  async insert(index: number, leaf: bigint): Promise<void> {
+    assertOrErr(index >= this.localTree.count, "index must be >= tree count");
+    while (this.localTree.count < index) {
+      this.localTree.insert(0n);
     }
 
-    await this.db.kv.putNumber(MERKLE_NEXT_BLOCK_TO_INDEX, latestBlock + 1);
-  }
-
-  async fetchNewLeaves(from: number, to: number): Promise<bigint[]> {
-    const insertions = await fetchInsertions(this.contract, from, to);
-    return insertions.map((insertion) => {
-      if (typeof insertion === "bigint") {
-        // it's a note commitment
-        return insertion;
-      } else {
-        // it's a note
-        return NoteTrait.toCommitment(insertion);
-      }
-    });
+    this.localTree.insert(leaf);
   }
 }