+++ packages/sdk/src/store/index.ts
@@ -1,4 +1,2 @@
 export { KVStore, DumpableKVStore, KV } from "./kvStore";
 export { InMemoryKVStore } from "./inMemoryKVStore";
-export { NotesDB } from "./notes";
-export { MerkleDB } from "./merkle";